requirements.txt
olefile
pyunpack
patool
python-magic